+++ admin/views/statuses/view.html.php
@@ -1,16 +1,16 @@
 <?php
 /**
- * @package      UserIdeas
+ * @package      Userideas
  * @subpackage   Component
  * @author       Todor Iliev
- * @copyright    Copyright (C) 2015 Todor Iliev <todor@itprism.com>. All rights reserved.
+ * @copyright    Copyright (C) 2016 Todor Iliev <todor@itprism.com>. All rights reserved.
  * @license      GNU General Public License version 3 or later; see LICENSE.txt
  */
 
 // no direct access
 defined('_JEXEC') or die;
 
-class UserIdeasViewStatuses extends JViewLegacy
+class UserideasViewStatuses extends JViewLegacy
 {
     /**
      * @var JDocumentHtml
@@ -32,17 +32,14 @@ class UserIdeasViewStatuses extends JViewLegacy
     protected $saveOrder;
     protected $saveOrderingUrl;
     protected $sortFields;
+    protected $canDo;
 
     protected $sidebar;
 
-    public function __construct($config)
-    {
-        parent::__construct($config);
-        $this->option = JFactory::getApplication()->input->get("option");
-    }
-
     public function display($tpl = null)
     {
+        $this->option = JFactory::getApplication()->input->get('option');
+        
         $this->state      = $this->get('State');
         $this->items      = $this->get('Items');
         $this->pagination = $this->get('Pagination');
@@ -66,7 +63,7 @@ class UserIdeasViewStatuses extends JViewLegacy
         // Prepare filters
         $this->listOrder = $this->escape($this->state->get('list.ordering'));
         $this->listDirn  = $this->escape($this->state->get('list.direction'));
-        $this->saveOrder = (strcmp($this->listOrder, 'a.ordering') != 0) ? false : true;
+        $this->saveOrder = (strcmp($this->listOrder, 'a.ordering') === 0);
 
         if ($this->saveOrder) {
             $this->saveOrderingUrl = 'index.php?option=' . $this->option . '&task=' . $this->getName() . '.saveOrderAjax&format=raw';
@@ -84,7 +81,7 @@ class UserIdeasViewStatuses extends JViewLegacy
      */
     protected function addSidebar()
     {
-        UserIdeasHelper::addSubmenu($this->getName());
+        UserideasHelper::addSubmenu($this->getName());
 
         $this->sidebar = JHtmlSidebar::render();
     }
@@ -97,16 +94,27 @@ class UserIdeasViewStatuses extends JViewLegacy
      */
     protected function addToolbar()
     {
-        // Set toolbar items for the page
+        $this->canDo = JHelperContent::getActions('com_userideas');
+
         JToolbarHelper::title(JText::_('COM_USERIDEAS_STATUSES_MANAGER'));
-        JToolbarHelper::addNew('status.add');
-        JToolbarHelper::editList('status.edit');
-        JToolbarHelper::divider();
-        JToolbarHelper::makeDefault('statuses.setDefault', 'COM_USERIDEAS_DEFAULT');
-        JToolbarHelper::divider();
-        JToolbarHelper::deleteList(JText::_("COM_USERIDEAS_DELETE_ITEMS_QUESTION"), "statuses.delete");
-        JToolbarHelper::divider();
-        JToolbarHelper::custom('statuses.backToDashboard', "dashboard", "", JText::_("COM_USERIDEAS_DASHBOARD"), false);
+
+        if ($this->canDo->get('core.create')) {
+            JToolbarHelper::addNew('status.add');
+        }
+
+        if ($this->canDo->get('core.edit')) {
+            JToolbarHelper::editList('status.edit');
+        }
+
+        if ($this->canDo->get('core.edit.state')) {
+            JToolbarHelper::makeDefault('statuses.setDefault', 'COM_USERIDEAS_DEFAULT');
+        }
+
+        if ($this->canDo->get('core.delete')) {
+            JToolbarHelper::deleteList(JText::_('COM_USERIDEAS_DELETE_ITEMS_QUESTION'), 'statuses.delete');
+        }
+
+        JToolbarHelper::custom('statuses.backToDashboard', 'dashboard', '', JText::_('COM_USERIDEAS_DASHBOARD'), false);
     }
 
     /**
